requirements.txt
absl-py==0.6.1
astor==0.7.1
Click==7.0
Cython==0.29.2
gast==0.2.1
grpcio==1.17.1
h5py==2.9.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.4
pandas==0.23.4
pkg-resources==0.0.0
protobuf==3.6.1
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
word2vec==0.10.2
